chore(filters): dry filters drawer into separate files
