+++ src/components/FiltersDrawer/index.ts
@@ -1,0 +1,1 @@
+export { FiltersDrawer as default } from './FiltersDrawer';
