+++ src/index.ts
@@ -30,6 +30,7 @@ export { default as FieldsetItems } from './components/FieldsetItems';
 export { default as FieldsetLegend } from './components/FieldsetLegend';
 export { default as Filters } from './components/Filters';
 export { default as FiltersCheckboxField } from './components/FiltersCheckboxField';
+export { default as FiltersDrawer } from './components/FiltersDrawer';
 export { default as Grid } from './components/Grid';
 export { default as GridItem } from './components/GridItem';
 export { default as Heading } from './components/Heading';
